requirements.txt
# Core dependencies
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0

# Optional for development
jupyter>=1.0.0
notebook>=6.4.0

# For data visualization (future use)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0 